refactor: 유저 정보 조회 api
